requirements.txt
docxtpl>=0.16.0
python-docx>=0.8.11
docxcompose>=1.4.0
Jinja2>=3.0.0
Pillow>=9.0.0
